🔖 (v0.3.0): Admin role integration (#10)
This commits updates the project with several improvements to configuration management, admin features, and homepage rendering.

- Hardcoded configuration values updated: The `main.py` file now uses `app.config` instead of hardcoded values for configuration.
- Admin users moved to environment variable in JSON format: The `config.py` file has been updated to read admin users from an environment variable in JSON format, replacing the hardcoded list.
- Homepage rendering updated based on user role: The `main.py` file now updates homepage rendering based on user roles.
- Split homepage into Jinja templates: The `templates/` directory now includes separate templates for different user roles.
- Admins can update requests for all users: A new feature in `main.py` allows admins to update requests for all users.
- New Admin template created: A `home.html` template has been added for the admin role.
- Admin list added to configuration file: The `config.py` file now includes a list of admin users.

## Commits

* 📝 (config.py): add a list of admin users to the configuration file

* 📝 (templates/home.html): create Admin template

* ✨ (main.py): introduce new functionality to allow admins to update requests for all users

* 🌐 (templates/): split homepage in jinja templates

In this new templates subdivision 'home.html' template is been extended by user, admin and unlogged pages.
In the home template there are 2 main parts to be extended:
 - scripts: js script (if needed)
 - body: html text

In this way user role and admin role 's html are splitted  and it's easy
to manage and access the code.

* 🌐 (main.py): update homepage rendering based on user role

* 📝 (admin.html): remove empty `<script>` tag in Admin template

* 🔧 (config.py): update the file to read admin users from an environment variable in JSON format instead of a hardcoded list

* 🔧 (main.py): update hardcoded configuration values to use app.config instead

+++ main.py
@@ -41,19 +41,19 @@ app = Flask(__name__)
 
 dbms = DBManager()
 
-app.logger.debug("Setting secret key")
-app.secret_key = '!secret'      # TODO: sceglierne una migliore !
-
 app.logger.debug("Loading configs from envs")
 app.config.from_object('config')
 
-# TODO: nascondere in .env ?
+app.logger.debug(f"Setting secret key: {app.config['SECRET_KEY']}")
+app.secret_key = app.config['SECRET_KEY']
+
+app.logger.debug(f"ADMIN USERS: {app.config['ADMIN_USERS']}")
+
 app.logger.debug("Loading OAuth configs")
-CONF_URL = 'http://127.0.0.1:8200/v1/identity/oidc/provider/default/.well-known/openid-configuration'
 oauth = OAuth(app)
 oauth.register(
     name='vault',
-    server_metadata_url=CONF_URL,
+    server_metadata_url=app.config['VAULT_CONF_URL'],
     api_base_url='http://localhost:8200/v1/',
     client_kwargs={
         'scope': 'openid default'
@@ -69,18 +69,40 @@ def homepage():
     app.logger.debug(f"/ User value: {user}")
 
     if user:
-        match request.method:
-            case 'POST':
-                dbms.add_request(user['contact']['email'])
-            case 'DELETE':
-                dbms.delete_request(user['contact']['email'])
-            case _:
-                print(request.method)
+        # ADMIN ROLE
+        if user['contact']['email'] in app.config['ADMIN_USERS']:
+            app.logger.debug(f"User {user['contact']['email']} is ADMIN")
+            user['admin'] = True
+            
+            if request.method == 'POST':
+                user_to_update = request.form['id']
+                app.logger.debug(f"UPDATING USER {user_to_update} REQUEST")
 
-        user['request_data'] = dbms.get_request_status(user['contact']['email'])
-        app.logger.debug(f"USER REQUEST DATA: {user['request_data']}")
+                dbms.update_request_status(user_to_update)
 
-    return render_template('home.html', user=user)
+            user['request_data'] = dbms.get_all_requests_status()
+            app.logger.debug(f"ADMIN REQUEST DATA: {user['request_data']}")
+
+            return render_template('users/admin.html', user=user)
+
+        # USER ROLE
+        else:
+            user['admin'] = False
+
+            match request.method:
+                case 'POST':
+                    dbms.add_request(user['contact']['email'])
+                case 'DELETE':
+                    dbms.delete_request(user['contact']['email'])
+                case _:
+                    print(request.method)
+
+            user['request_data'] = dbms.get_request_status(user['contact']['email'])
+            app.logger.debug(f"USER REQUEST DATA: {user['request_data']}")
+
+            return render_template('users/user.html', user=user)
+
+    return render_template('users/unlogged.html', user=user)
 
 
 @app.route('/login')
